test: every fixture lives in the root conftests and a gate test fails if a nested conftest defines one, so pytest 9.1.1 (pytest-dev/pytest#14971) cannot drop a fixture; the pin below 9 is lifted, the directory-scoped autouse fixtures gate on the test path, and the straddle-overlap override no longer requests monkeypatch so fixture teardown order matches the pre-migration tree
