+++ tests/conftest.py
@@ -3,16 +3,20 @@ import os
 import re
 import secrets
 import sys
-from collections.abc import AsyncGenerator, Callable
+import uuid
+from collections.abc import AsyncGenerator, Callable, Generator, Iterator
 from importlib.machinery import ModuleSpec
 from pathlib import Path
 from types import FrameType
-from typing import Any
+from typing import TYPE_CHECKING, Any, cast
+from unittest.mock import AsyncMock, MagicMock, patch
 from urllib.parse import parse_qsl
 
 import pytest
+from guard_agent.models import AgentConfig, SecurityEvent, SecurityMetric
 from pytest import TempPathFactory
 
+from guard_core._utils import body_reader
 from guard_core._utils import detection_scan as _detection_scan_module
 from guard_core.core.events import logfire_handler as _logfire_handler_module
 from guard_core.core.events import otel_handler as _otel_handler_module
@@ -40,6 +44,9 @@ from guard_core.sync.handlers.ratelimit_handler import (
     rate_limit_handler as _sync_rate_limit_handler,
 )
 
+if TYPE_CHECKING:
+    from tests.live_smoke.driver import ScenarioContext, Stack
+
 _suspatterns_module._legacy_detection_warned = True
 
 IPINFO_TOKEN = os.getenv("IPINFO_TOKEN") or "test_token"
@@ -49,6 +56,13 @@ REDIS_PREFIX = f"{_REDIS_PREFIX_BASE}{os.getpid()}:{secrets.token_hex(4)}:"
 GUARD_TESTS_GC_PER_TEST = os.getenv("GUARD_TESTS_GC_PER_TEST") == "1"
 _MAX_TEST_SECONDS_ENV = "GUARD_TESTS_MAX_TEST_SECONDS"
 _TEST_NODE_DURATIONS: dict[str, float] = {}
+
+_TESTS_DIR = Path(__file__).parent
+_TEST_AGENT_DIR = _TESTS_DIR / "test_agent"
+_TEST_SUS_PATTERNS_DIR = _TESTS_DIR / "test_sus_patterns"
+_TEST_CLOUD_IPS_DIR = _TESTS_DIR / "test_cloud_ips"
+_TEST_UTILS_DIR = _TESTS_DIR / "test_utils"
+_CLOUD_IP_REDIS_PREFIX = f"test:guard_core_cloud_ip_isolation:{uuid.uuid4().hex}:"
 
 
 @pytest.fixture(autouse=True)
@@ -436,6 +450,350 @@ def _collect_garbage_after_test() -> Any:
     yield
     if GUARD_TESTS_GC_PER_TEST:
         gc.collect()
+
+
+@pytest.fixture
+def mock_guard_agent() -> Generator[Any, Any, Any]:
+    import sys
+    import types
+
+    mock_guard_agent_module = types.ModuleType("guard_agent")
+    guard_agent_ns = cast(Any, mock_guard_agent_module)
+    guard_agent_ns.SecurityEvent = SecurityEvent
+    guard_agent_ns.SecurityMetric = SecurityMetric
+    guard_agent_ns.AgentConfig = AgentConfig
+
+    mock_models_module = types.ModuleType("guard_agent.models")
+    guard_agent_models_ns = cast(Any, mock_models_module)
+    guard_agent_models_ns.SecurityEvent = SecurityEvent
+    guard_agent_models_ns.SecurityMetric = SecurityMetric
+    guard_agent_models_ns.AgentConfig = AgentConfig
+    guard_agent_ns.models = mock_models_module
+
+    mock_agent_handler = AsyncMock()
+    mock_guard_agent_func = MagicMock(return_value=mock_agent_handler)
+    guard_agent_ns.guard_agent = mock_guard_agent_func
+
+    original_modules = {}
+    modules_to_mock = [
+        "guard_agent",
+        "guard_agent.models",
+    ]
+
+    for module_name in modules_to_mock:
+        if module_name in sys.modules:
+            original_modules[module_name] = sys.modules[module_name]
+
+    sys.modules["guard_agent"] = mock_guard_agent_module
+    sys.modules["guard_agent.models"] = mock_models_module
+
+    with (
+        patch(
+            "guard_core.handlers.behavior_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.cloud_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.dynamic_rule_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.decorators.base.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.ipban_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.ipinfo_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.ratelimit_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.redis_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.handlers.suspatterns_handler.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.utils.SecurityEvent",
+            SecurityEvent,
+            create=True,
+        ),
+        patch(
+            "guard_core.models.AgentConfig",
+            AgentConfig,
+            create=True,
+        ),
+    ):
+        try:
+            yield mock_guard_agent_module
+        finally:
+            for module_name in modules_to_mock:
+                if module_name in original_modules:
+                    sys.modules[module_name] = original_modules[module_name]
+                elif module_name in sys.modules:  # pragma: no cover
+                    del sys.modules[module_name]
+
+
+@pytest.fixture(autouse=True)
+def mock_dependencies(request: pytest.FixtureRequest) -> Generator[Any, Any, Any]:
+    if not request.path.is_relative_to(_TEST_AGENT_DIR):
+        yield
+        return
+
+    request.getfixturevalue("mock_guard_agent")
+    with (
+        patch(
+            "guard_core.handlers.redis_handler.RedisManager.initialize",
+            new_callable=AsyncMock,
+        ),
+        patch(
+            "guard_core.handlers.ipinfo_handler.IPInfoManager.__new__"
+        ) as mock_ipinfo,
+        patch("guard_core.handlers.cloud_handler.CloudManager.__new__") as mock_cloud,
+    ):
+        mock_ipinfo_instance = MagicMock()
+        mock_ipinfo.return_value = mock_ipinfo_instance
+
+        mock_cloud_instance = MagicMock()
+        mock_cloud.return_value = mock_cloud_instance
+        yield
+
+
+@pytest.fixture
+def config() -> SecurityConfig:
+    return SecurityConfig(
+        enable_agent=True,
+        agent_api_key="test-api-key",
+        agent_endpoint="http://test.example.com",
+        enable_dynamic_rules=True,
+        dynamic_rule_interval=60,
+        enable_penetration_detection=True,
+        enable_ip_banning=True,
+        enable_rate_limiting=True,
+        rate_limit=100,
+        rate_limit_window=60,
+        auto_ban_threshold=5,
+    )
+
+
+@pytest.fixture
+def mock_agent_handler() -> AsyncMock:
+    handler = AsyncMock()
+    handler.get_dynamic_rules = AsyncMock(return_value=None)
+    handler.send_event = AsyncMock()
+    return handler
+
+
+@pytest.fixture
+def mock_redis_handler() -> AsyncMock:
+    return AsyncMock()
+
+
+@pytest.fixture
+def security_config_with_detection() -> SecurityConfig:
+    return SecurityConfig(
+        detection_compiler_timeout=2.0,
+        detection_max_content_length=10000,
+        detection_preserve_attack_patterns=True,
+        detection_semantic_threshold=0.7,
+        detection_anomaly_threshold=3.0,
+        detection_slow_pattern_threshold=0.1,
+        detection_monitor_history_size=1000,
+        detection_max_tracked_patterns=1000,
+    )
+
+
+@pytest.fixture
+async def sus_patterns_manager_with_detection(
+    security_config_with_detection: SecurityConfig,
+) -> AsyncGenerator[SusPatternsManager, None]:
+    original_instance = SusPatternsManager._instance
+    original_config = SusPatternsManager._config
+
+    SusPatternsManager._instance = None
+    SusPatternsManager._config = None
+
+    manager = SusPatternsManager(security_config_with_detection)
+
+    yield manager
+
+    await manager.reset()
+
+    SusPatternsManager._instance = original_instance
+    SusPatternsManager._config = original_config
+
+
+@pytest.fixture(autouse=True)
+async def reset_sus_patterns(
+    request: pytest.FixtureRequest,
+) -> AsyncGenerator[None, None]:
+    if not request.path.is_relative_to(_TEST_SUS_PATTERNS_DIR):
+        yield
+        return
+
+    original_instance = SusPatternsManager._instance
+    original_config = SusPatternsManager._config
+
+    original_patterns = None
+    original_custom_patterns: set[str] = set()
+    if original_instance:
+        original_patterns = original_instance.patterns.copy()
+        original_custom_patterns = original_instance.custom_patterns.copy()
+
+    yield
+
+    if SusPatternsManager._instance:
+        await SusPatternsManager._instance.reset()
+
+    SusPatternsManager._instance = original_instance
+    SusPatternsManager._config = original_config
+
+    if original_instance and original_patterns:
+        original_instance.patterns = original_patterns
+        original_instance.custom_patterns = original_custom_patterns
+
+
+@pytest.fixture
+def security_config_redis_isolated_prefix(ipinfo_db_path: Path) -> SecurityConfig:
+    return SecurityConfig(
+        redis_url=REDIS_URL,
+        redis_prefix=_CLOUD_IP_REDIS_PREFIX,
+        whitelist=["127.0.0.1"],
+        blacklist=["192.168.1.1"],
+        blocked_user_agents=[r"badbot"],
+        auto_ban_threshold=3,
+        auto_ban_duration=300,
+        custom_log_file="test_log.log",
+        custom_error_responses={
+            403: "Custom Forbidden",
+            429: "Custom Too Many Requests",
+        },
+        enable_cors=True,
+        cors_allow_origins=["https://example.com"],
+        cors_allow_methods=["GET", "POST"],
+        cors_allow_headers=["*"],
+        cors_allow_credentials=True,
+        cors_expose_headers=["X-Custom-Header"],
+        cors_max_age=600,
+    )
+
+
+async def _flush_cloud_ip_redis_namespace() -> None:
+    config = SecurityConfig(redis_url=REDIS_URL, redis_prefix=_CLOUD_IP_REDIS_PREFIX)
+    redis_handler = RedisManager(config)
+    await redis_handler.initialize()
+    try:
+        await redis_handler.delete_pattern("*")
+    except Exception:
+        pass
+    finally:
+        await redis_handler.close()
+
+
+@pytest.fixture(autouse=True)
+async def cloud_ip_redis_isolation(
+    request: pytest.FixtureRequest,
+) -> AsyncGenerator[None, None]:
+    if not request.path.is_relative_to(_TEST_CLOUD_IPS_DIR):
+        yield
+        return
+    await _flush_cloud_ip_redis_namespace()
+    yield
+    await _flush_cloud_ip_redis_namespace()
+
+
+@pytest.fixture(autouse=True)
+def _zero_straddle_overlap_by_default(
+    request: pytest.FixtureRequest,
+) -> Generator[Any, Any, Any]:
+    if not request.path.is_relative_to(_TEST_UTILS_DIR):
+        yield
+        return
+
+    async def _zero_overlap() -> int:
+        return 0
+
+    with patch.object(body_reader, "_straddle_overlap_bytes", _zero_overlap):
+        yield
+
+
+def _check_live_smoke_preconditions(stack_dir: Path) -> None:
+    baseline = os.environ.get("LIVE_SMOKE_BASELINE")
+    wheels = list((stack_dir / "wheels").glob("*.whl"))
+    if not baseline and not wheels:
+        pytest.fail(
+            "no guard-core wheel in tests/live_smoke/stack/wheels/; run "
+            "`uv build --wheel --out-dir tests/live_smoke/stack/wheels` first, "
+            "or set LIVE_SMOKE_BASELINE=guard-core==X.Y.Z to test a released "
+            "version instead. See `make live-smoke`.",
+            pytrace=False,
+        )
+    if not (stack_dir / "app" / "security.py").is_file():
+        pytest.fail(
+            "tests/live_smoke/stack/app/ is missing; run "
+            "`python tests/live_smoke/fetch_example_app.py` then "
+            "`python tests/live_smoke/patch_example_config.py` first. "
+            "See `make live-smoke`.",
+            pytrace=False,
+        )
+
+
+@pytest.fixture(scope="session")
+def stack() -> "Iterator[Stack]":
+    from tests.live_smoke.driver import STACK_DIR, Stack
+
+    _check_live_smoke_preconditions(STACK_DIR)
+    instance = Stack()
+    instance.up()
+    try:
+        yield instance
+    finally:
+        instance.down()
+
+
+@pytest.fixture(scope="session")
+def scenario_context(stack: "Stack") -> "Iterator[ScenarioContext]":
+    from tests.live_smoke.driver import (
+        ScenarioContext,
+        make_agent_client,
+        make_http_client,
+        make_otlp_client,
+        make_redis_client,
+    )
+
+    client = make_http_client()
+    agent = make_agent_client()
+    redis_client = make_redis_client()
+    otlp = make_otlp_client()
+    try:
+        yield ScenarioContext(
+            stack=stack, client=client, agent=agent, redis=redis_client, otlp=otlp
+        )
+    finally:
+        client.close()
+        agent.close()
+        redis_client.close()
+        otlp.close()
 
 
 _GUARD_AGENT_FINDER_ATTR = "_guard_core_agent_import_finder"
