+++ tests/live_smoke/conftest.py
@@ -2,20 +2,9 @@ from __future__ import annotations
 
 import importlib
 import os
-from collections.abc import Iterator
 from pathlib import Path
 
 import pytest
-
-from tests.live_smoke.driver import (
-    STACK_DIR,
-    ScenarioContext,
-    Stack,
-    make_agent_client,
-    make_http_client,
-    make_otlp_client,
-    make_redis_client,
-)
 
 _HERE = Path(__file__).resolve().parent
 
@@ -55,52 +44,3 @@ def pytest_collection_modifyitems(
 
     ours.sort(key=lambda item: item.path.name == "test_completeness.py")
     items[:] = others + ours
-
-
-def _check_preconditions() -> None:
-    baseline = os.environ.get("LIVE_SMOKE_BASELINE")
-    wheels = list((STACK_DIR / "wheels").glob("*.whl"))
-    if not baseline and not wheels:
-        pytest.fail(
-            "no guard-core wheel in tests/live_smoke/stack/wheels/; run "
-            "`uv build --wheel --out-dir tests/live_smoke/stack/wheels` first, "
-            "or set LIVE_SMOKE_BASELINE=guard-core==X.Y.Z to test a released "
-            "version instead. See `make live-smoke`.",
-            pytrace=False,
-        )
-    if not (STACK_DIR / "app" / "security.py").is_file():
-        pytest.fail(
-            "tests/live_smoke/stack/app/ is missing; run "
-            "`python tests/live_smoke/fetch_example_app.py` then "
-            "`python tests/live_smoke/patch_example_config.py` first. "
-            "See `make live-smoke`.",
-            pytrace=False,
-        )
-
-
-@pytest.fixture(scope="session")
-def stack() -> Iterator[Stack]:
-    _check_preconditions()
-    instance = Stack()
-    instance.up()
-    try:
-        yield instance
-    finally:
-        instance.down()
-
-
-@pytest.fixture(scope="session")
-def scenario_context(stack: Stack) -> Iterator[ScenarioContext]:
-    client = make_http_client()
-    agent = make_agent_client()
-    redis_client = make_redis_client()
-    otlp = make_otlp_client()
-    try:
-        yield ScenarioContext(
-            stack=stack, client=client, agent=agent, redis=redis_client, otlp=otlp
-        )
-    finally:
-        client.close()
-        agent.close()
-        redis_client.close()
-        otlp.close()
