+++ tests/test_cloud_ips/test_cloud_ips.py
@@ -955,7 +955,9 @@ async def test_fetch_azure_ip_ranges_sizes_timeout_from_remaining_budget(  # asy
     assert second_attempt_timeout.kwargs["total"] == 5.0
 
 
-async def test_cloud_ip_redis_caching(security_config_redis: SecurityConfig) -> None:
+async def test_cloud_ip_redis_caching(
+    security_config_redis_isolated_prefix: SecurityConfig,
+) -> None:
     with (
         patch(
             "guard_core.handlers.cloud_handler.fetch_aws_ip_ranges",
@@ -989,7 +991,7 @@ async def test_cloud_ip_redis_caching(security_config_redis: SecurityConfig) -> 
     ):
         mock_aws.return_value = {ipaddress.IPv4Network("192.168.0.0/24")}
 
-        redis_handler = RedisManager(security_config_redis)
+        redis_handler = RedisManager(security_config_redis_isolated_prefix)
         await redis_handler.initialize()
 
         await cloud_handler.initialize_redis(redis_handler)
@@ -1018,11 +1020,11 @@ async def test_cloud_ip_redis_caching(security_config_redis: SecurityConfig) -> 
 
 
 async def test_cloud_ip_redis_cache_hit(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
     import json as _json
 
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
 
     await redis_handler.set_key("cloud_ip_v2", "AWS", _json.dumps(["192.168.0.0/24"]))
@@ -1040,7 +1042,7 @@ async def test_cloud_ip_redis_cache_hit(
 
 
 async def test_cloud_ip_redis_sync_async(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
     with (
         patch(
@@ -1080,7 +1082,7 @@ async def test_cloud_ip_redis_sync_async(
         await cloud_handler.refresh()
         assert cloud_handler.is_cloud_ip("192.168.0.1", {"AWS"})
 
-        redis_handler = RedisManager(security_config_redis)
+        redis_handler = RedisManager(security_config_redis_isolated_prefix)
         await redis_handler.initialize()
         await cloud_handler.initialize_redis(redis_handler)
 
@@ -1092,7 +1094,7 @@ async def test_cloud_ip_redis_sync_async(
 
 
 async def test_cloud_ip_redis_error_handling(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
     with patch(
         "guard_core.handlers.cloud_handler.fetch_aws_ip_ranges",
@@ -1100,7 +1102,7 @@ async def test_cloud_ip_redis_error_handling(
     ) as mock_aws:
         mock_aws.return_value = {ipaddress.IPv4Network("192.168.0.0/24")}
 
-        redis_handler = RedisManager(security_config_redis)
+        redis_handler = RedisManager(security_config_redis_isolated_prefix)
         await redis_handler.initialize()
 
         await redis_handler.delete("cloud_ranges_v2", "AWS")
@@ -1355,9 +1357,9 @@ async def test_new_providers_wired_into_refresh_async_store_path() -> None:
 
 
 async def test_new_providers_wired_into_refresh_via_redis_handler(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
     await redis_handler.delete("cloud_ranges", "DigitalOcean")
     await redis_handler.delete("cloud_ranges", "Linode")
@@ -1442,9 +1444,9 @@ async def test_refresh_via_redis_handler_falls_back_when_redis_missing() -> None
 
 
 async def test_refresh_via_redis_handler_records_empty_on_fetch_error(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
     await redis_handler.delete("cloud_ranges", "AWS")
 
@@ -1538,7 +1540,7 @@ async def test_fetch_gcp_ip_ranges_skips_unknown_prefix_keys(
 
 
 async def test_initialize_redis_replaces_in_memory_store(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
     from guard_core.handlers.cloud_ip_stores import (
         InMemoryCloudIpStore,
@@ -1549,7 +1551,7 @@ async def test_initialize_redis_replaces_in_memory_store(
     cloud_handler.redis_handler = None
     cloud_handler.ip_ranges = {}
 
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
 
     with (
@@ -1601,14 +1603,14 @@ async def test_initialize_agent_records_handler() -> None:
 
 
 async def test_initialize_redis_keeps_existing_redis_store(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
     from guard_core.handlers.cloud_ip_stores import (
         InMemoryCloudIpStore,
         RedisCloudIpStore,
     )
 
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
 
     cloud_handler.set_store(RedisCloudIpStore(redis_handler))
@@ -1657,9 +1659,9 @@ async def test_initialize_redis_keeps_existing_redis_store(
 
 
 async def test_refresh_via_redis_handler_handles_empty_fetch(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
     await redis_handler.delete("cloud_ranges", "AWS")
 
@@ -1704,9 +1706,9 @@ def test_cloud_manager_returns_existing_singleton() -> None:
 
 
 async def test_refresh_via_redis_handler_keeps_existing_provider_state(
-    security_config_redis: SecurityConfig,
+    security_config_redis_isolated_prefix: SecurityConfig,
 ) -> None:
-    redis_handler = RedisManager(security_config_redis)
+    redis_handler = RedisManager(security_config_redis_isolated_prefix)
     await redis_handler.initialize()
     await redis_handler.delete("cloud_ranges", "AWS")
